+++ sound-mind-studio/tests/test_tool_palette_controller.h
@@ -1,0 +1,19 @@
+#pragma once
+
+#include <QObject>
+
+class ToolPaletteControllerTest : public QObject {
+    Q_OBJECT
+
+private slots:
+    void freshControllerHasNoSelection();
+    void setProjectDoesNotCrash();
+    void beginAndCancelEachToolDoesNotCrash();
+    void undoAndRedoOnEmptyHistoryDoNotCrash();
+    void setPathPlacesSmoothNodesDoesNotCrash();
+    void setGridSnappingDoesNotCrash();
+    void pasteIntoWithNothingCopiedReturnsNullopt();
+    void copyAndCutWithNoSelectionDoNotCrash();
+    void fillWithNoSelectionDoesNotCrash();
+    void contentChangedAggregatesAPaintStroke();
+};